requirements.txt
pyzmq
opencv-python
pyaudio
numpy
PyQt5